+++ tests/testthat/test-remotes.R
@@ -21,33 +21,43 @@ test_that("we can parse a variety of remotes", {
   record <- renv_remotes_resolve("kevinushey/skeleton")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.1")
+  expect_equal(record$RemoteRef, "master")
+  expect_equal(record$RemoteSha, "958296dbbbf7f1d82f7f5dd1b121c7558604809f")
 
   # by commit
   record <- renv_remotes_resolve("kevinushey/skeleton@209c4e48e505e545ad7ab915904d983b5ab83b93")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.0")
+  expect_equal(record$RemoteSha, "209c4e48e505e545ad7ab915904d983b5ab83b93")
 
   # by branch
   record <- renv_remotes_resolve("kevinushey/skeleton@feature/version-bump")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.2")
+  expect_equal(record$RemoteRef, "feature/version-bump")
+  expect_equal(record$RemoteSha, "86b5737411d3c6a6927dfcccd2c15a69284659fe")
 
   # by PR
   record <- renv_remotes_resolve("kevinushey/skeleton#1")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.2")
+  expect_equal(record$RemoteSha, "86b5737411d3c6a6927dfcccd2c15a69284659fe")
 
   # bitbucket
   record <- renv_remotes_resolve("bitbucket::kevinushey/skeleton")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.1")
   expect_equal(record$RemoteHost, "api.bitbucket.org/2.0")
+  expect_equal(record$RemoteRef, "master")
+  expect_equal(record$RemoteSha, "958296dbbbf7f1d82f7f5dd1b121c7558604809f")
 
   # gitlab
   record <- renv_remotes_resolve("gitlab::kevinushey/skeleton")
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.1")
   expect_equal(record$RemoteHost, "gitlab.com")
+  expect_equal(record$RemoteRef, "master")
+  expect_equal(record$RemoteSha, "958296dbbbf7f1d82f7f5dd1b121c7558604809f")
 
   # error
   expect_error(renv_remotes_resolve("can't parse this"))